add search bar component
