+++ resources/js/Pages/Home/Components/SearchBar.vue
@@ -17,28 +17,63 @@
             />
             <span
                 class="clear-btn"
-                v-if="searchQuery.length > 0"
-                @click="searchQuery = ''"
+                v-if="searchQuery.length > 0 && !loading"
+                @click="clearSearch"
             >×</span>
         </div>
     </div>
 </template>
 
 <script>
+import { ref, watch } from 'vue';
+import { debounce } from 'lodash-es';
+
 export default {
-    props: {
-        modelValue: String
-    },
-    emits: ['update:modelValue', 'toggle-sidebar', 'input-focus'],
-    computed: {
-        searchQuery: {
-            get() {
-                return this.modelValue;
-            },
-            set(value) {
-                this.$emit('update:modelValue', value);
+    props: {},
+    emits: ['toggle-sidebar', 'input-focus', 'results', 'loading'],
+    setup(props, { emit }) {
+        const searchQuery = ref('');
+        const loading = ref(false);
+        const csrfToken = ref(document.querySelector('meta[name="csrf-token"]')?.content || '');
+
+        const clearSearch = () => {
+            searchQuery.value = '';
+        };
+
+        watch(searchQuery, debounce((newVal) => {
+            if (newVal.trim().length === 0) {
+                emit('results', []);
+                return;
             }
-        }
+
+            loading.value = true;
+            emit('loading', true);
+
+            fetch(route('search-users', { username: newVal }), {
+                headers: {
+                    'Accept': 'application/json',
+                    'X-CSRF-TOKEN': csrfToken.value,
+                    'X-Requested-With': 'XMLHttpRequest'
+                }
+            })
+                .then(res => res.json())
+                .then(data => {
+                    emit('results', data);
+                })
+                .catch(() => {
+                    emit('results', []);
+                })
+                .finally(() => {
+                    loading.value = false;
+                    emit('loading', false);
+                });
+        }, 300));
+
+        return {
+            searchQuery,
+            loading,
+            clearSearch
+        };
     }
 };
 </script>
